Preserve Voicemail Settings During FreePBX Extension Modifications via GQL API

+++ Core.class.php
@@ -2400,11 +2400,13 @@ class Core extends FreePBX_Helpers implements BMO  {
 			}
 
 			//voicemail symlink
-			$spooldir = $this->config->get('ASTSPOOLDIR');
-			$account = preg_replace("/\D/","",$account);
-			if(freepbx_trim ($account) !== "" && file_exists($spooldir."/voicemail/device/".$account)) {
-				exec("rm -f ".escapeshellarg($spooldir."/voicemail/device/".$account));
-			}
+            if (!$editmode) {
+                $spooldir = $this->config->get('ASTSPOOLDIR');
+                $account = preg_replace("/\D/", "", $account);
+                if (freepbx_trim($account) !== "" && file_exists($spooldir . "/voicemail/device/" . $account)) {
+                    exec("rm -f " . escapeshellarg($spooldir . "/voicemail/device/" . $account));
+                }
+            }
 		} else {
 			die_freepbx("Cannot connect to Asterisk Manager with ".$this->config->get("AMPMGRUSER")."/".$this->config->get("AMPMGRPASS"));
 		}
